backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
openai
supabase
